removed header component
