+++ login-workflow/src/components/WorkflowCard/WorkflowCard.tsx
@@ -1,34 +1,17 @@
 import React from 'react';
-import { ImageBackground, ImageBackgroundProps, ImageSourcePropType, StyleSheet } from 'react-native';
-import { Card, CardActionsProps, CardProps, CardTitleProps } from 'react-native-paper';
-import { WorkflowCardInstructionProps } from './WorkflowCardInstructions';
+import { ImageBackground, StyleSheet } from 'react-native';
+import { Card, CardProps } from 'react-native-paper';
 import { useScreenDimensions } from '../../hooks/useScreenDimensions';
 import { ExtendedTheme, useExtendedTheme } from '@brightlayer-ui/react-native-themes';
 import { EdgeInsets, useSafeAreaInsets } from 'react-native-safe-area-context';
 import { Spinner } from '../Spinner';
-import defaultImage from '../../assets/images/background.png';
-import { WorkflowCardHeader } from './WorkflowCardHeader';
+import { WorkflowCardBody } from './WorkflowCardBody';
+import { WorkflowCardBaseProps } from './WorkflowCard.types';
+
+const defaultImage = require('../../assets/images/background.png');
 
 const MAX_CARD_HEIGHT = 730;
 const MAX_CARD_WIDTH = 450;
-/**
- * Component that renders the workflow card that is used for all screen components.
- *
- * @param loading boolean value for isLoading
- * @param backgroundImage to display card background
- *
- * @category Component
- */
-export type WorkflowCardBaseProps = ImageBackgroundProps & {
-    /**
-     * If true, a blocking progress spinner will be displayed over the card
-     */
-    loading?: boolean;
-    /**
-     * A custom background to render behind the card
-     */
-    backgroundImage?: ImageSourcePropType;
-};
 
 const makeStyles = ({
     insets,
@@ -55,21 +38,19 @@ const makeStyles = ({
         },
     });
 
-export type WorkflowCardHeaderProps = CardTitleProps;
-
-export type WorkflowCardActionsProps = CardActionsProps;
-
-export type WorkflowCardProps = {
-    WorkflowCardBaseProps?: WorkflowCardBaseProps;
-    WorkflowCardHeaderProps?: WorkflowCardHeaderProps;
-    WorkflowCardInstructionProps?: WorkflowCardInstructionProps;
-    WorkflowCardActionsProps?: WorkflowCardActionsProps;
-};
-
 function hasWorkflowCardHeaderRecursive(children: any): boolean {
-    return React.Children.toArray(children).some((child) => (child as JSX.Element).type === WorkflowCardHeader);
+    // @todo replace WorkflowCardBody with WorkflowCardHeader component
+    return React.Children.toArray(children).some((child) => (child as JSX.Element).type === WorkflowCardBody);
 }
 
+/**
+ * Component that renders the workflow card that is used for all screen components.
+ *
+ * @param loading boolean value for isLoading
+ * @param backgroundImage to display card background
+ *
+ * @category Component
+ */
 export const WorkflowCard: React.FC<WorkflowCardBaseProps> = (props) => {
     const { loading, backgroundImage, children, style, ...otherImageProps } = props;
     const theme = useExtendedTheme();
