+++ login-workflow/src/components/WorkflowCard/WorkflowCard.types.ts
@@ -1,7 +1,33 @@
 import { DotStepperVariant } from '@brightlayer-ui/react-native-components';
-import { CardActionsProps } from 'react-native-paper';
+import { ImageBackgroundProps, ImageSourcePropType, ViewProps } from 'react-native';
+import { CardTitleProps, TextProps } from 'react-native-paper';
 
-export type WorkflowCardActionsProps = Omit<CardActionsProps, 'children'> & {
+export type WorkflowCardBaseProps = ImageBackgroundProps & {
+    /**
+     * If true, a blocking progress spinner will be displayed over the card
+     */
+    loading?: boolean;
+    /**
+     * A custom background to render behind the card
+     */
+    backgroundImage?: ImageSourcePropType;
+};
+
+export type WorkflowCardHeaderProps = CardTitleProps;
+
+export type WorkflowCardInstructionProps = Omit<TextProps<'bodyLarge'>, 'children' | 'theme' | 'variant'> & {
+    /**
+     * The text to display as instructions
+     */
+    instructions?: string | JSX.Element;
+    /**
+     * Whether or not to show a divider below the instructions
+     * @default true
+     */
+    divider?: boolean;
+};
+
+export type WorkflowCardActionsProps = ViewProps & {
     /**
      * True to display a divider above workflow card action buttons
      * @default true
@@ -76,4 +102,11 @@ export type WorkflowCardActionsProps = Omit<CardActionsProps, 'children'> & {
      * @default: 'dots'
      */
     stepperVariant?: DotStepperVariant;
+};
+
+export type WorkflowCardProps = {
+    WorkflowCardBaseProps?: WorkflowCardBaseProps;
+    WorkflowCardHeaderProps?: WorkflowCardHeaderProps;
+    WorkflowCardInstructionProps?: WorkflowCardInstructionProps;
+    WorkflowCardActionsProps?: WorkflowCardActionsProps;
 };
