+++ login-workflow/src/components/WorkflowCard/index.tsx
@@ -2,4 +2,3 @@ export * from './WorkflowCard';
 export * from './WorkflowCardBody';
 export * from './WorkflowCardInstructions';
 export * from './WorkflowCardActions';
-export * from './WorkflowCardHeader';
